+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Tags are now highlighted and returned to original value if a write fails
 * Applicable tags are now marked as stale when realtime updating fails
 
+### Minor Changes
+
+* Added option to disable strict checking for allowed tag name characters
+  
 ## v1.1.9
 
 ### Major Changes